Regenerate lexer/parser with ANTLR 4.9.2
I couldn't find the procedure used to generate the previous files
documented anywhere, so I made this up which seems to produce a
reasonable `git diff`:

  antlr4 -Dlanguage=Python3 sphinx_a4doc/syntax/ANTLRv4Lexer.g4
  antlr4 -Dlanguage=Python3 -visitor sphinx_a4doc/syntax/ANTLRv4Parser.g4
  mv sphinx_a4doc/syntax/ANTLRv4Lexer.py sphinx_a4doc/syntax/ANTLRv4Parser.py sphinx_a4doc/syntax/ANTLRv4ParserListener.py sphinx_a4doc/syntax/ANTLRv4ParserVisitor.py sphinx_a4doc/syntax/gen/syntax/
  rm sphinx_a4doc/syntax/*.interp sphinx_a4doc/syntax/ANTLRv4Parser.tokens

I then edited `sphinx_a4doc/syntax/gen/syntax/ANTLRv4Lexer.py` manually
to remove `from .LexerAdaptor import LexerAdaptor` imports, which I
couldn't find a way to suppress.  (There's probably a better way to do
this.)

I chose version 4.9.2 partly because it's the version in Debian testing,
but also because 4.10 makes an incompatible changes to the format of ATN
serialization data, which seemed more invasive.  You may want to do that
at some point, but for now I thought it was best to get things working
on Python 3.13 in a more minimal way.

Fixes: #20